Langchain/requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4

wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2

numpy
pandas
ipykernel

langchain-groq
langchain-objectbox